+++ src/LamaPon/Graphics/D3D12RenderServices.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <memory>
+
+namespace LamaPon
+{
+    class D3D12Backend;
+    class GraphicsRenderServices;
+
+    [[nodiscard]] std::unique_ptr<GraphicsRenderServices>
+        CreateD3D12GraphicsRenderServices(D3D12Backend& backend);
+}